Fix pagination of nominations
Previously, all nominations were displayed on the first page, with blank
pages following. This change introduces an islice for the iterable that
should match the expected pagination strategy.

It should be noted that Py3.12 introduces itertools.batched, which
approximates this; however, this commit does not presume to change
Py3.12 to the minimum necessary version.