Add requirements and docker-compose
The requirements.txt file should codify the precise package requirements
of this project. This allows environments to be repeatably set up.

The docker-compose file provides a straightforward way to launch the
service with minimal configuration, in a repeatable and testable way.

Both together should lower the barrier to development and testing.